Converted commands into their own files. * sp!dm * sp!dmmute * sp!help * sp!info * sp!invite * sp!lang * sp!mute * sp!neofetch * sp!ping * sp!set * Converted mention reply into a seperate file. * Converted trigger and insult handling into a seperate file.
